fix(infowindow): replace useDeepCompareEffect with useEffect (#786)
Due to the destructuring assignment of infowindowOptions, the useDeepCompareEffect was used to avoid re-assigning options even if they didn't change.

This, however, caused serious problems with other dependencies of those effects that could contain cycles and large object-structures to be compared.

Instead of `useDeepCompareEffect` this switches to using regular effects and deep memoization for the infowindowOptions.